refactor: move buy crypto request models to buy crypto workflow
