--- imx-core-sdk-android/src/main/java/com/immutable/sdk/model/GetSignedMoonpayRequest.kt
@@ -1,6 +1,0 @@
-package com.immutable.sdk.model
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-internal data class GetSignedMoonpayRequest(val request: String)
